Move content_type_from_file_name to utils.
